+++ CMakeLists.txt
@@ -107,6 +107,7 @@ add_executable(alicia-server
         src/server/messenger/MessengerDirector.cpp
         src/server/race/RaceDirector.cpp
         src/server/ranch/RanchDirector.cpp
+        src/server/room/Room.cpp
         src/server/system/ChatSystem.cpp
         src/server/system/InfractionSystem.cpp
         src/server/system/ItemSystem.cpp
