+++ include/server/system/RoomSystem.hpp
@@ -20,6 +20,8 @@
 #ifndef ROOMREGISTRY_HPP
 #define ROOMREGISTRY_HPP
 
+#include "server/room/Room.hpp"
+
 #include <libserver/data/DataDefinitions.hpp>
 
 #include <cstdint>
@@ -31,87 +33,6 @@
 
 namespace server
 {
-
-class Room
-{
-public:
-  enum class GameMode
-  {
-    Speed = 1, Magic = 2, Guild, Tutorial = 6
-  };
-
-  enum class TeamMode
-  {
-    FFA = 1,
-    Team = 2,
-    Single = 3
-  };
-
-  class Player
-  {
-  public:
-    enum class Team
-    {
-      Solo, Red, Blue
-    };
-
-    bool ToggleReady();
-    void SetReady(bool ready);
-    [[nodiscard]] bool IsReady() const;
-    void SetTeam(Team team);
-    [[nodiscard]] Team GetTeam() const;
-  private:
-    bool _isReady = false;
-    Team _team = Team::Solo;
-  };
-
-  struct Details
-  {
-    std::string name;
-    std::string password;
-    uint16_t missionId{};
-    uint16_t courseId{};
-    uint32_t maxPlayerCount{};
-    GameMode gameMode{};
-    TeamMode teamMode{};
-    uint8_t npcDifficulty{};
-    uint8_t skillBracket{};
-  };
-
-  struct Snapshot
-  {
-    uint32_t uid;
-    Details details;
-    size_t playerCount;
-    bool isPlaying;
-  };
-
-  explicit Room(uint32_t uid);
-
-  [[nodiscard]] bool IsRoomFull() const;
-  bool QueuePlayer(data::Uid characterUid);
-  bool DequeuePlayer(data::Uid characterUid);
-  bool AddPlayer(data::Uid characterUid);
-  void RemovePlayer(data::Uid characterUid);
-  [[nodiscard]] Player& GetPlayer(data::Uid characterUid);
-
-  void SetRoomPlaying(bool isPlaying);
-
-  [[nodiscard]] uint32_t GetUid() const;
-  [[nodiscard]] bool IsRoomPlaying() const;
-  [[nodiscard]] size_t GetPlayerCount() const;
-
-  [[nodiscard]] Details& GetRoomDetails();
-  [[nodiscard]] Snapshot GetRoomSnapshot() const;
-  [[nodiscard]] std::unordered_map<data::Uid, Player>& GetPlayers();
-
-private:
-  Details _details;
-  uint32_t _uid{};
-  std::unordered_set<data::Uid> _queuedPlayers;
-  std::unordered_map<data::Uid, Player> _players;
-  bool _roomIsPlaying{};
-};
 
 class RoomSystem
 {
